weekend contacts are now functional.  The daily update is functional but not perfect.  Fixed the memory fragmentation issues.
outstanding issues:
there is a bug in the reproduction number counting methods.  For some reason the kernel isn't adding the counts up properly
all seasonal generations are listed as zero.  possibly related to the above, but maybe not
The rebuild_infected_array method is still really rough, some quick cleanup should be possible.  Or switch to global status arrays and rebuild the infected arrays daily